Añadir Activity para ver receta. Configurar el ViewPager2 incluid en un navhostfragment mediante un grafo de navegación. Tres framentos por cada pestaña que controla el tablayout y que gestiona el ViewPager2 mediante TabLayoutMediator

+++ app/src/main/java/org/overdrive/recetasdigitales/model/dao/IngredienteDAO.java
@@ -20,5 +20,5 @@ public interface IngredienteDAO {
     // Recuperar ingredientes de una receta
     @Transaction
     @Query("SELECT * FROM recetas WHERE idReceta = :id")
-    LiveData<RecetaIngredientes> getRecetaConIngredientes(int id);
+    LiveData<RecetaIngredientes> getRecetaConIngredientes(long id);
 }
